refactor: changed folder structure for land and landTunnl tests
